requirements.txt
Flask
flask-cors
requests
python-dotenv
pandas
gunicorn
